Move Enumerator back into NonEmptyEnumerable.cs
Drop the partial-class split — it only existed to move the nested
Enumerator into a separate file. Since callers never name the enumerator
type directly (foreach uses it implicitly), the single-file shape is
simpler without any real cost.
