+++ src/StrongTypes/Collections/NonEmptyEnumerable.cs
@@ -80,7 +80,7 @@ public static class NonEmptyEnumerable
 [CollectionBuilder(typeof(NonEmptyEnumerable), nameof(NonEmptyEnumerable.Create))]
 [DebuggerTypeProxy(typeof(NonEmptyEnumerableDebugView<>))]
 [DebuggerDisplay("Count = {Count}")]
-public sealed partial class NonEmptyEnumerable<T> : INonEmptyEnumerable<T>, ICollection<T>, IEquatable<NonEmptyEnumerable<T>>
+public sealed class NonEmptyEnumerable<T> : INonEmptyEnumerable<T>, ICollection<T>, IEquatable<NonEmptyEnumerable<T>>
 {
     // Owned buffer — factories copy inputs, callers can never mutate us from outside.
     private readonly T[] _values;
@@ -148,6 +148,27 @@ public sealed partial class NonEmptyEnumerable<T> : INonEmptyEnumerable<T>, ICol
     bool ICollection<T>.Remove(T item) => throw new NotSupportedException("NonEmptyEnumerable is read-only.");
 
     #endregion
+
+    public struct Enumerator : IEnumerator<T>
+    {
+        private readonly T[] _values;
+        private int _index;
+
+        internal Enumerator(T[] values)
+        {
+            _values = values;
+            _index = -1;
+        }
+
+        public readonly T Current => _values[_index];
+        readonly object? IEnumerator.Current => Current;
+
+        public bool MoveNext() => ++_index < _values.Length;
+
+        public void Reset() => _index = -1;
+
+        public readonly void Dispose() { }
+    }
 
     #region Equality
